Template change and partial inclusion of contests.
My first attempt at working with a Roll20 template--yay! I began by noticing that the templates weren't using some of the values passed to them. This evolved into combining all the template functionality into one. Next, I tried to incorporate more of the contest functionality into the sheet. It's not really complete, but hopefully this works as a starting point. As I mentioned, this is my first time working with these, so there may be some bugs (especially with "allprops", which may be duplicating output).